2019/11/08 16：54

+++ src/main/java/com/kp/domain/User.java
@@ -1,11 +1,13 @@
 package com.kp.domain;
 
+import java.math.BigInteger;
+
 public class User {
 
     private Integer user_id;
     private String user_name;
     private String user_password;
-    private Integer user_phone;
+    private BigInteger user_phone;
     private Integer user_status;
 
     @Override
@@ -43,11 +45,11 @@ public class User {
         this.user_password = user_password;
     }
 
-    public Integer getUser_phone() {
+    public BigInteger getUser_phone() {
         return user_phone;
     }
 
-    public void setUser_phone(Integer user_phone) {
+    public void setUser_phone(BigInteger user_phone) {
         this.user_phone = user_phone;
     }
